Moved the AuthStateListener Activity in the com.example.musicmap package
